Change score back to an int

+++ src/main/java/com/google/sps/utils/UserUtils.java
@@ -32,7 +32,7 @@ import com.google.sps.utils.QuizTimingPropertiesUtils;
 
 public final class UserUtils {
     public static final String SESSION_ID_COOKIE_NAME = "SessionID";
-    public static final long ADDED_POINTS = 20L;
+    public static final int ADDED_POINTS = 20;
     private static final Logger log = Logger.getLogger(UserUtils.class.getName());
 
     /**
@@ -144,9 +144,9 @@ public final class UserUtils {
   /**
     adds a specified number of points to the user's points
   */
-  public static void addPoints(Entity userEntity, long numPoints, DatastoreService datastore){
+  public static void addPoints(Entity userEntity, int numPoints, DatastoreService datastore){
     try {
-      userEntity.setProperty("score", ((Number) userEntity.getProperty("score")).longValue() + numPoints);
+      userEntity.setProperty("score", ((Number) userEntity.getProperty("score")).intValue() + numPoints);
     } catch (NullPointerException e) {
       userEntity.setProperty("score", numPoints);                
     }
